Spun out the PhysicalEntity class into an ActorEntity class. This will be where logic specific to things that move go - namely weapon and inventory information, when it comes to that.
Cleaned up the InputHandler code a bit, so an entity doesn't have to know
about their input handler. I'm still not really happy with how this stuff
works, but for now it's sufficient. Hopefully we can refactor it in the
near future to something that feels a bit cleaner.

Added an interface for weapons within the game. The next step is to add
some cool weapons. Coming soon: awesomeness.