New Features!
Added double jump - the player can now jump twice! I reduced jump height
though, so it's more reasonable. Fixes #3.

Fixed the trigger system that I broke. Fixes #23.

Added simple enemies - the most basic AI walks in a straight line,
changing direction when it touches a wall. Refs #16.

Added a simple weapon/projectile system. As is par for the course,
everything explodes when things get touched. Fixes #13. I also changed the
interact button's behavior so that it acts as an attack button if there is
nothing to be interacted with. Lots more can be done here, refs #27.

Did a bit of code cleanup, but lots of work remains to be done here...